# Make sure the optimizer does not remove the code