My results:

```
running 7 tests
test array_concat        ... bench:          46 ns/iter (+/- 1)
test array_join          ... bench:          19 ns/iter (+/- 0)
test format_macro        ... bench:          99 ns/iter (+/- 7)
test mut_string_push_str ... bench:          44 ns/iter (+/- 1)
test string_from_plus_op ... bench:          43 ns/iter (+/- 1)
test to_owned_plus_op    ... bench:          50 ns/iter (+/- 38)
test to_string_plus_op   ... bench:          72 ns/iter (+/- 3)

test result: ok. 0 passed; 0 failed; 0 ignored; 7 measured
```