@badboy 2016-02-29
https://github.com/rust-lang/rust/blob/504ca6f4221b5745c5d72902c1c30415845afa26/src/libtest/lib.rs#L1178-L1190