@hoodie 2016-02-29
awesome, array joining still beets everything, even array concatenation, for whatever reason